+++ internal-api/src/main/java/datadog/trace/util/LongHashingUtils.java
@@ -48,7 +48,7 @@ public final class LongHashingUtils {
     return hash(intHash(obj0), intHash(obj1));
   }
 
-  public static final long hash(int hash0, int hash1) {
+  static final long hash(int hash0, int hash1) {
     return 31L * hash0 + hash1;
   }
 
@@ -60,7 +60,7 @@ public final class LongHashingUtils {
     return hash(intHash(obj0), intHash(obj1), intHash(obj2));
   }
 
-  public static final long hash(int hash0, int hash1, int hash2) {
+  static final long hash(int hash0, int hash1, int hash2) {
     // DQH - Micro-optimizing, 31L * 31L will constant fold
     // Since there are multiple execution ports for load & store,
     // this will make good use of the core.
@@ -71,7 +71,7 @@ public final class LongHashingUtils {
     return hash(intHash(obj0), intHash(obj1), intHash(obj2), intHash(obj3));
   }
 
-  public static final long hash(int hash0, int hash1, int hash2, int hash3) {
+  static final long hash(int hash0, int hash1, int hash2, int hash3) {
     // DQH - Micro-optimizing, 31L * 31L will constant fold
     // Since there are multiple execution ports for load & store,
     // this will make good use of the core.
@@ -82,7 +82,7 @@ public final class LongHashingUtils {
     return hash(intHash(obj0), intHash(obj1), intHash(obj2), intHash(obj3), intHash(obj4));
   }
 
-  public static final long hash(int hash0, int hash1, int hash2, int hash3, int hash4) {
+  static final long hash(int hash0, int hash1, int hash2, int hash3, int hash4) {
     // DQH - Micro-optimizing, 31L * 31L will constant fold
     // Since there are multiple execution ports for load & store,
     // this will make good use of the core.
