+++ internal-api/src/test/java/datadog/trace/util/HashtableTest.java
@@ -1,243 +1,23 @@
 package datadog.trace.util;
 
+import static datadog.trace.util.HashtableTestEntries.CollidingKey;
+import static datadog.trace.util.HashtableTestEntries.CollidingKeyEntry;
+import static datadog.trace.util.HashtableTestEntries.StringIntEntry;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import datadog.trace.util.Hashtable.BucketIterator;
 import datadog.trace.util.Hashtable.MutatingBucketIterator;
 import datadog.trace.util.Hashtable.Support;
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Set;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 class HashtableTest {
-
-  // ============ D1 ============
-
-  @Nested
-  class D1Tests {
-
-    @Test
-    void emptyTableLookupReturnsNull() {
-      Hashtable.D1<String, StringIntEntry> table = new Hashtable.D1<>(8);
-      assertNull(table.get("missing"));
-      assertEquals(0, table.size());
-    }
-
-    @Test
-    void insertedEntryIsRetrievable() {
-      Hashtable.D1<String, StringIntEntry> table = new Hashtable.D1<>(8);
-      StringIntEntry e = new StringIntEntry("foo", 1);
-      table.insert(e);
-      assertEquals(1, table.size());
-      assertSame(e, table.get("foo"));
-    }
-
-    @Test
-    void multipleInsertsRetrievableSeparately() {
-      Hashtable.D1<String, StringIntEntry> table = new Hashtable.D1<>(16);
-      StringIntEntry a = new StringIntEntry("alpha", 1);
-      StringIntEntry b = new StringIntEntry("beta", 2);
-      StringIntEntry c = new StringIntEntry("gamma", 3);
-      table.insert(a);
-      table.insert(b);
-      table.insert(c);
-      assertEquals(3, table.size());
-      assertSame(a, table.get("alpha"));
-      assertSame(b, table.get("beta"));
-      assertSame(c, table.get("gamma"));
-    }
-
-    @Test
-    void inPlaceMutationVisibleViaSubsequentGet() {
-      Hashtable.D1<String, StringIntEntry> table = new Hashtable.D1<>(8);
-      table.insert(new StringIntEntry("counter", 0));
-      for (int i = 0; i < 10; i++) {
-        StringIntEntry e = table.get("counter");
-        e.value++;
-      }
-      assertEquals(10, table.get("counter").value);
-    }
-
-    @Test
-    void removeUnlinksEntryAndDecrementsSize() {
-      Hashtable.D1<String, StringIntEntry> table = new Hashtable.D1<>(8);
-      table.insert(new StringIntEntry("a", 1));
-      table.insert(new StringIntEntry("b", 2));
-      assertEquals(2, table.size());
-
-      StringIntEntry removed = table.remove("a");
-      assertNotNull(removed);
-      assertEquals("a", removed.key);
-      assertEquals(1, table.size());
-      assertNull(table.get("a"));
-      assertNotNull(table.get("b"));
-    }
-
-    @Test
-    void removeNonexistentReturnsNullAndDoesNotChangeSize() {
-      Hashtable.D1<String, StringIntEntry> table = new Hashtable.D1<>(8);
-      table.insert(new StringIntEntry("a", 1));
-      assertNull(table.remove("nope"));
-      assertEquals(1, table.size());
-    }
-
-    @Test
-    void insertOrReplaceReturnsPriorEntryOrNullOnInsert() {
-      Hashtable.D1<String, StringIntEntry> table = new Hashtable.D1<>(8);
-      StringIntEntry first = new StringIntEntry("k", 1);
-      assertNull(table.insertOrReplace(first), "fresh insert returns null");
-      assertEquals(1, table.size());
-
-      StringIntEntry second = new StringIntEntry("k", 2);
-      assertSame(first, table.insertOrReplace(second), "replace returns the prior entry");
-      assertEquals(1, table.size());
-      assertSame(second, table.get("k"), "new entry visible after replace");
-    }
-
-    @Test
-    void clearEmptiesTheTable() {
-      Hashtable.D1<String, StringIntEntry> table = new Hashtable.D1<>(8);
-      table.insert(new StringIntEntry("a", 1));
-      table.insert(new StringIntEntry("b", 2));
-      table.clear();
-      assertEquals(0, table.size());
-      assertNull(table.get("a"));
-      // Reinsertion works after clear
-      table.insert(new StringIntEntry("a", 99));
-      assertEquals(99, table.get("a").value);
-    }
-
-    @Test
-    void forEachVisitsEveryInsertedEntry() {
-      Hashtable.D1<String, StringIntEntry> table = new Hashtable.D1<>(8);
-      table.insert(new StringIntEntry("a", 1));
-      table.insert(new StringIntEntry("b", 2));
-      table.insert(new StringIntEntry("c", 3));
-      Map<String, Integer> seen = new HashMap<>();
-      table.forEach(e -> seen.put(e.key, e.value));
-      assertEquals(3, seen.size());
-      assertEquals(1, seen.get("a"));
-      assertEquals(2, seen.get("b"));
-      assertEquals(3, seen.get("c"));
-    }
-
-    @Test
-    void nullKeyIsPermittedAndDistinctFromAbsent() {
-      Hashtable.D1<String, StringIntEntry> table = new Hashtable.D1<>(8);
-      assertNull(table.get(null));
-      StringIntEntry nullKeyed = new StringIntEntry(null, 7);
-      table.insert(nullKeyed);
-      assertSame(nullKeyed, table.get(null));
-      assertEquals(1, table.size());
-      assertSame(nullKeyed, table.remove(null));
-      assertEquals(0, table.size());
-    }
-
-    @Test
-    void hashCollisionsResolveByEquality() {
-      // Force two distinct keys with the same hashCode -- the chain must still distinguish them
-      // via matches().
-      Hashtable.D1<CollidingKey, CollidingKeyEntry> table = new Hashtable.D1<>(4);
-      CollidingKey k1 = new CollidingKey("first", 17);
-      CollidingKey k2 = new CollidingKey("second", 17);
-      CollidingKeyEntry e1 = new CollidingKeyEntry(k1, 100);
-      CollidingKeyEntry e2 = new CollidingKeyEntry(k2, 200);
-      table.insert(e1);
-      table.insert(e2);
-      assertEquals(2, table.size());
-      assertSame(e1, table.get(k1));
-      assertSame(e2, table.get(k2));
-    }
-
-    @Test
-    void hashCollisionsThenRemoveLeavesOtherIntact() {
-      Hashtable.D1<CollidingKey, CollidingKeyEntry> table = new Hashtable.D1<>(4);
-      CollidingKey k1 = new CollidingKey("first", 17);
-      CollidingKey k2 = new CollidingKey("second", 17);
-      CollidingKey k3 = new CollidingKey("third", 17);
-      table.insert(new CollidingKeyEntry(k1, 1));
-      table.insert(new CollidingKeyEntry(k2, 2));
-      table.insert(new CollidingKeyEntry(k3, 3));
-      table.remove(k2);
-      assertEquals(2, table.size());
-      assertNotNull(table.get(k1));
-      assertNull(table.get(k2));
-      assertNotNull(table.get(k3));
-    }
-  }
-
-  // ============ D2 ============
-
-  @Nested
-  class D2Tests {
-
-    @Test
-    void pairKeysParticipateInIdentity() {
-      Hashtable.D2<String, Integer, PairEntry> table = new Hashtable.D2<>(8);
-      PairEntry ab = new PairEntry("a", 1, 100);
-      PairEntry ac = new PairEntry("a", 2, 200);
-      PairEntry bb = new PairEntry("b", 1, 300);
-      table.insert(ab);
-      table.insert(ac);
-      table.insert(bb);
-      assertEquals(3, table.size());
-      assertSame(ab, table.get("a", 1));
-      assertSame(ac, table.get("a", 2));
-      assertSame(bb, table.get("b", 1));
-      assertNull(table.get("a", 3));
-    }
-
-    @Test
-    void removePairUnlinks() {
-      Hashtable.D2<String, Integer, PairEntry> table = new Hashtable.D2<>(8);
-      PairEntry ab = new PairEntry("a", 1, 100);
-      PairEntry ac = new PairEntry("a", 2, 200);
-      table.insert(ab);
-      table.insert(ac);
-      assertSame(ab, table.remove("a", 1));
-      assertEquals(1, table.size());
-      assertNull(table.get("a", 1));
-      assertSame(ac, table.get("a", 2));
-    }
-
-    @Test
-    void insertOrReplaceMatchesOnBothKeys() {
-      Hashtable.D2<String, Integer, PairEntry> table = new Hashtable.D2<>(8);
-      PairEntry first = new PairEntry("k", 7, 1);
-      assertNull(table.insertOrReplace(first));
-      PairEntry second = new PairEntry("k", 7, 2);
-      assertSame(first, table.insertOrReplace(second));
-      // Different second-key: should insert new, not replace
-      PairEntry third = new PairEntry("k", 8, 3);
-      assertNull(table.insertOrReplace(third));
-      assertEquals(2, table.size());
-    }
-
-    @Test
-    void forEachVisitsBothPairs() {
-      Hashtable.D2<String, Integer, PairEntry> table = new Hashtable.D2<>(8);
-      table.insert(new PairEntry("a", 1, 100));
-      table.insert(new PairEntry("b", 2, 200));
-      Set<String> seen = new HashSet<>();
-      table.forEach(e -> seen.add(e.key1 + ":" + e.key2));
-      assertEquals(2, seen.size());
-      assertTrue(seen.contains("a:1"));
-      assertTrue(seen.contains("b:2"));
-    }
-  }
 
   // ============ Support ============
 
@@ -374,7 +154,9 @@ class HashtableTest {
       // of the three keys are still retrievable.)
       int found = 0;
       for (CollidingKey k : new CollidingKey[] {k1, k2, k3}) {
-        if (table.get(k) != null) found++;
+        if (table.get(k) != null) {
+          found++;
+        }
       }
       assertEquals(2, found);
     }
@@ -411,8 +193,6 @@ class HashtableTest {
     }
   }
 
-  // ============ test helpers ============
-
   /** Reach into a D1 table's bucket array via reflection -- only needed by iterator tests. */
   private static Hashtable.Entry[] extractBuckets(Hashtable.D1<?, ?> table) {
     try {
@@ -423,68 +203,4 @@ class HashtableTest {
       throw new RuntimeException(e);
     }
   }
-
-  /** Sort comparator used by tests that want deterministic visit order. */
-  @SuppressWarnings("unused")
-  private static final Comparator<StringIntEntry> BY_KEY = Comparator.comparing(e -> e.key);
-
-  private static final class StringIntEntry extends Hashtable.D1.Entry<String> {
-    int value;
-
-    StringIntEntry(String key, int value) {
-      super(key);
-      this.value = value;
-    }
-  }
-
-  /** Key whose hashCode is fully controllable, to force chain collisions deterministically. */
-  private static final class CollidingKey {
-    final String label;
-    final int hash;
-
-    CollidingKey(String label, int hash) {
-      this.label = label;
-      this.hash = hash;
-    }
-
-    @Override
-    public int hashCode() {
-      return hash;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (!(o instanceof CollidingKey)) return false;
-      CollidingKey that = (CollidingKey) o;
-      return hash == that.hash && label.equals(that.label);
-    }
-
-    @Override
-    public String toString() {
-      return "CollidingKey(" + label + ", " + hash + ")";
-    }
-  }
-
-  private static final class CollidingKeyEntry extends Hashtable.D1.Entry<CollidingKey> {
-    int value;
-
-    CollidingKeyEntry(CollidingKey key, int value) {
-      super(key);
-      this.value = value;
-    }
-  }
-
-  private static final class PairEntry extends Hashtable.D2.Entry<String, Integer> {
-    int value;
-
-    PairEntry(String key1, Integer key2, int value) {
-      super(key1, key2);
-      this.value = value;
-    }
-  }
-
-  // Imports kept narrow but List is referenced in test helpers below; this keeps the import warning
-  // quiet.
-  @SuppressWarnings("unused")
-  private static final List<Object> UNUSED = new ArrayList<>();
 }
